Fix too many hooks reported in list_hooks_in_dir
list_hooks_in_dir should report exactly the hooks that would be run with
run_hooks. But the find command in list_hooks_in_dir does not limit itself to
the scheme "folder_name/hook_name" but recurses infinitely through sub
directories reporting all executable files. If your hooks are stored itself in
a git repository then consequently also the files in .git/hooks are reported.

This change sets the search depth exactly to two, so that the correct scripts
are reported, e.g. :
    ~/.git_hooks/commit-msg/script1

but not e.g. :

    ~/.git_hook/.git/hooks/commit-msg
    ~/.git_hooks/script2
    ~/.git_hooks/commit-msg/dir2/script3

Signed-off-by: Benjamin C Meyer <ben@meyerhome.net>