Add the ability to remind users when committing to any repo without git-hooks.
It is easy to forget to install git-hooks in a repository you are using.
A commit with no error message could both mean you don't have git-hooks
installed and it can mean that you do have it installed, but your
commit was perfect and there was no errors.

The new option --installglobal creates a default .git template
that is used by all future cloned repositories (and git init'd)
that remind users that the repository doesn't have git-hooks
installed.

A number of projects I have been involved with used git-hooks,
but founds its weakest point was that eventually someone would
re-clone a repo and make commits with errors before turning
git-hooks back on.  Another common problem was when we had
multiple git repositories and git-hooks was only installed in
some of them, but not all

While it would be nice to be able to use --instalglobal to
automatically turn on git-hooks on every future repo eventually
some git repository on github would include a hook that would
email <secret stuff> back home and all you would have to do
would be to clone it to activate it which would be a "bad thing",
but it can pester you without harm.

Signed-off-by: Benjamin C Meyer <ben@meyerhome.net>